Refactor: Home screen divided into view and view-model.
